Conflicts:
	Debug/SDL_Game_Framework.exe
	Debug/SDL_Game_Framework.ilk
	Debug/SDL_Game_Framework.pdb
	SDL_Game_Framework.sdf
	SDL_Game_Framework.v12.suo
	SDL_Game_Framework/Debug/Level.obj
	SDL_Game_Framework/Debug/Level_One.obj
	SDL_Game_Framework/Debug/Level_Three.obj
	SDL_Game_Framework/Debug/Level_Two.obj
	SDL_Game_Framework/Debug/Play_Screen.obj
	SDL_Game_Framework/Debug/SDL_Game.8E28E0D2.tlog/CL.read.1.tlog
	SDL_Game_Framework/Debug/SDL_Game.8E28E0D2.tlog/CL.write.1.tlog
	SDL_Game_Framework/Debug/SDL_Game.8E28E0D2.tlog/link.write.1.tlog
	SDL_Game_Framework/Debug/SDL_Game_Framework.log
	SDL_Game_Framework/Debug/vc120.idb
	SDL_Game_Framework/Debug/vc120.pdb

+++ SDL_Game_Framework/AW_Sprite_Interface.cpp
@@ -145,7 +145,7 @@ void AW_Sprite_Interface::Randomize_Position(int xPos, int width, int yPos, int 
 {
 	int rand_x = rand() % width;
 	int rand_y = rand() % height;
-	set_world_position(xPos + rand_x, yPos + rand_y);
+	set_world_position((float)xPos + (float)rand_x, (float)yPos + (float)rand_y);
 }
 
 void AW_Sprite_Interface::Debug()
